Move SendEmail into the EmailJob to keep functionality together There wasn't much left except one method when refactoring was done.
